fixBug:result quiz return NaN
