+++ app/(auth)/auth/layout.tsx
@@ -1,4 +1,3 @@
-import { ToastWrapper } from "@/components";
 import { Metadata } from "next";
 import { FC, ReactNode } from "react";
