+++ app/(auth)/auth/login/page.tsx
@@ -1,6 +1,6 @@
+import { LoginModule } from "@/components/pages";
 import { NextPage } from "next";
 import { ReactElement } from "react";
-import { LoginModule } from "./_modules/login-module";
 
 const Login: NextPage = (): ReactElement => {
   return <LoginModule />;
