+++ app/(auth)/auth/register/page.tsx
@@ -1,9 +1,9 @@
+import { RegisterModule } from "@/components/pages";
 import { NextPage } from "next";
 import { ReactElement } from "react";
-import { RegisterModule } from "./_modules/register";
 
-const RegisterPage:NextPage = ():ReactElement => {
-    return <RegisterModule/>    
+const RegisterPage: NextPage = (): ReactElement => {
+  return <RegisterModule />;
 };
 
 export default RegisterPage;
