+++ app/(protected)/playground/layout.tsx
@@ -1,16 +1,14 @@
 import { FC, ReactNode } from "react";
-import { Header } from "./_modules/header";
+import { Header } from "@/components/molecules";
 import { Metadata } from "next";
 
 interface LayoutProps {
   children: ReactNode;
 }
-
 export const metadata: Metadata = {
   title: "Quiztify  | Playground",
   description: "Your Playground",
 };
-
 const Layout: FC<LayoutProps> = ({ children }) => {
   return (
     <main className="container min-h-screen flex flex-col">
