+++ components/atoms/button/button.tsx
@@ -1,7 +1,8 @@
 import { forwardRef } from "react";
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "@/utilities";
+
 import { Icon } from "@iconify/react/dist/iconify.js";
+import { cn } from "@/utilities/cn";
 
 const buttonVariants = cva(
   "inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:pointer-events-none disabled:opacity-50",
