+++ components/molecules/index.ts
@@ -1,1 +1,2 @@
-export * from './fields';
+export * from "./fields";
+export * from "./playground-header";
