+++ components/molecules/playground-header/header.tsx
@@ -11,7 +11,7 @@ export const Header: FC = (): ReactElement => {
   const toggleDropdown = () => setDropdownOpen(!dropdownOpen);
 
   return (
-    <header className="w-full md:w-4/5 mx-auto rounded-xl mt-4 bg-indigo-700 text-white shadow-lg">
+    <header className="w-full md:w-4/5 mx-auto rounded-xl mt-4 bg-gradient-to-r from-indigo-500 to-fuchsia-500 text-white shadow-lg">
       <div className="flex items-center justify-between w-full px-8 py-4">
         <h1 className="text-sm md:text-lg font-semibold">
           Welcome to your Playground
@@ -30,7 +30,7 @@ export const Header: FC = (): ReactElement => {
           </div>
 
           {dropdownOpen && (
-            <div className="absolute right-0 mt-2 w-40 bg-indigo-800 rounded-lg shadow-lg hover:bg-indigo-800/90">
+            <div className="absolute right-0 mt-2 w-40 bg-gradient-to-r from-indigo-500 to-fuchsia-500 rounded-lg shadow-lg ">
               <Button
                 variant="ghost"
                 className="w-full text-left  text-sm text-white transition-all"
