+++ components/pages/index.ts
@@ -1,0 +1,5 @@
+export * from "./register";
+export * from "./login-module";
+export * from "./quiz-result";
+export * from "./playground-module";
+export * from "./home";
