+++ components/pages/login-module.tsx
@@ -1,6 +1,6 @@
 "use client";
-import { signIn } from "next-auth/react";
 import React, { FC, ReactElement, useState } from "react";
+import { signIn } from "next-auth/react";
 import { Button, ControlledFieldText, ToastWrapper } from "@/components";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
