+++ prisma/schema.prisma
@@ -12,7 +12,6 @@ model User {
   name          String?
   email         String    @unique
   password      String
-  emailVerified DateTime?
   image         String?
   createdAt     DateTime  @default(now())
   updatedAt     DateTime  @updatedAt
