+++ utilities/cn.ts
@@ -1,0 +1,5 @@
+import clsx, { ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
